+++ src/citra_qt/debugger/graphics_framebuffer.cpp
@@ -10,6 +10,8 @@
 #include <QSpinBox>
 
 #include "core/hw/gpu.h"
+#include "core/memory.h"
+
 #include "video_core/color.h"
 #include "video_core/pica.h"
 #include "video_core/utils.h"
