+++ src/core/hle/function_wrappers.h
@@ -7,7 +7,7 @@
 #include "common/common_types.h"
 
 #include "core/arm/arm_interface.h"
-#include "core/mem_map.h"
+#include "core/memory.h"
 #include "core/hle/hle.h"
 
 namespace HLE {
@@ -109,7 +109,7 @@ template<ResultCode func(s64*, u32, void*, s32)> void Wrap(){
 
 template<ResultCode func(u32*, const char*)> void Wrap() {
     u32 param_1 = 0;
-    u32 retval = func(&param_1, Memory::GetCharPointer(PARAM(1))).raw;
+    u32 retval = func(&param_1, (char*)Memory::GetPointer(PARAM(1))).raw;
     Core::g_app_core->SetReg(1, param_1);
     FuncReturn(retval);
 }
@@ -163,7 +163,7 @@ template<void func(s64)> void Wrap() {
 }
 
 template<void func(const char*)> void Wrap() {
-    func(Memory::GetCharPointer(PARAM(0)));
+    func((char*)Memory::GetPointer(PARAM(0)));
 }
 
 #undef PARAM
