+++ src/core/hle/kernel/session.h
@@ -6,7 +6,7 @@
 
 #include "core/hle/kernel/kernel.h"
 #include "core/hle/kernel/thread.h"
-#include "core/mem_map.h"
+#include "core/memory.h"
 
 namespace Kernel {
