+++ src/core/hle/kernel/thread.h
@@ -12,7 +12,6 @@
 #include "common/common_types.h"
 
 #include "core/core.h"
-#include "core/mem_map.h"
 
 #include "core/hle/kernel/kernel.h"
 #include "core/hle/result.h"
