+++ src/core/hle/shared_page.cpp
@@ -8,7 +8,7 @@
 #include "common/common_funcs.h"
 
 #include "core/core.h"
-#include "core/mem_map.h"
+#include "core/memory.h"
 #include "core/hle/config_mem.h"
 #include "core/hle/shared_page.h"
