+++ src/core/loader/loader.cpp
@@ -13,7 +13,6 @@
 #include "core/loader/3dsx.h"
 #include "core/loader/elf.h"
 #include "core/loader/ncch.h"
-#include "core/mem_map.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
