+++ src/core/loader/ncch.cpp
@@ -12,7 +12,7 @@
 
 #include "core/hle/kernel/kernel.h"
 #include "core/loader/ncch.h"
-#include "core/mem_map.h"
+#include "core/memory.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // Loader namespace
